Update and align parsing of "\r\nOK\r\n> "
The SPI layer removes all unwanted \r\n that are starting a new line.
Should be better handler at higher layer, but this is the situation now.

This results in 2 possibles cases when reading data:
- if no data WIFI chipset returns "\r\n\r\nOK\r\n> ", SPI layer then removes
the double \r\n which results in AT/ISM layer receving "OK\r\n> ". This is
the no data case.
- if there is data: "\r\n<Data>\r\n\r\nOK\r\n> ", then SPI layer removes
the first \r\n, and receives DATA the "\r\nOK\r\n> " that needs to be
removed.

This is applied to both recv and check_recv_status