Fix max handling in BufferedSpi::read
this was not visible before because we were reading 1200 bytes,
but now we just read as much as possible.

in case max read is requested (which is all we use today), then
we should apply the 2 above filters, whatever len value.

So if
max == 0 => we always filter,
if max !=o, then we just need to chack that len is < max.

( (max == 0) || ((max != 0) && (max < len)) )
is equivalent to
( (max == 0) || (max < len) )