Detect connection closed or lost cases
As mentioned in module spec, when reading data possible cases are:
- Data received
or
- Empty content
or
- Response=”-1”
Any one of the following reasons will cause a “-1” response:
1. Connection to the network (AP) has been lost
2. Other side has closed the connection
3. Other side has lost it connection to the network

In case of "-1" content, the check_recv_status will return a negative
value, which means connection has been lost or closed. This information
needs to be passed to higher layers for proper action.